2025-09-22-rag_word_puzzle/rag_magic/pyproject.toml
[project]
name = "rag-magic"
version = "0.1.0"
description = "A CLI tool for RAG (Retrieval-Augmented Generation) operations with PostgreSQL and vector embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "tiktoken>=0.5.0",
    "numpy>=1.24.0",
    "pgvector>=0.2.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-community>=0.0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
rag-magic = "rag_magic.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_magic"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
